+++ include/documentmodel.h
@@ -162,6 +162,8 @@ typedef struct tagEDIT_CONFIGURATION {
 	int				hscroll;				// scroll n columns
 	int				vscroll;				// scroll n Lines
 	int				scroll_dy;				// scroll on dist dy to screen
+	void*			documentType;			// the document type descriptor from which this configuration was created or NULL for "default" document types not known to PKS-EDIT
+											// i.e. - no file name pattern matches etc...
 } EDIT_CONFIGURATION;
 
 /*--------------------------------------------------------------------------